pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "RGBMatrixDriver"
dynamic = ["version"]
description = "A driver middleware for LED matrix software compatible with rpi-rgb-led-matrix"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Tyler Porter", email = "tyler.b.porter@gmail.com" },
]
keywords = [
    "LED",
    "LED matrix",
    "RPI",
    "matrix",
    "raspberry",
    "raspberry pi",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Other/Nonlisted Topic",
]
dependencies = [
    "keyboard",
    "numpy>=1.2.0",
    "Pillow>=10.0.1",
    "RGBMatrixEmulator"
]

[project.urls]
Homepage = "https://github.com/ty-porter/RGBMatrixDriver"

[tool.hatch.version]
path = "RGBMatrixDriver/version.py"

[tool.hatch.build.targets.wheel.shared-data]
LICENSE = "docs/LICENSE"
"README.md" = "docs/README.md"

[tool.hatch.build.targets.sdist]
include = [
    "/rgbmatrixdriver",
]
